Fix: Make balance field double

+++ database/migrations/1767842764410_checking_accounts.ts
@@ -10,7 +10,7 @@ export default class extends BaseSchema {
       table.integer('agency_id').unsigned().references('agencies.id').onDelete('CASCADE')
 
       table.string('account_number', 17)
-      table.integer('balance')
+      table.double('balance')
       table.string('password', 255)
 
       /**
